Safari-609.1.11.1/LayoutTests/inspector/dom/shadowRootType-expected.txt

Test for DOMNode.shadowRootType.


== Running test suite: DOMNode.shadowRootType
-- Running test case: ShadowRootType.UserAgent
PASS: DOMNode has a single shadow root.
PASS: DOMNode has UserAgent shadow root type.
PASS: DOMNode is in a UserAgent shadow tree.

-- Running test case: ShadowRootType.Closed
PASS: DOMNode has a single shadow root.
PASS: DOMNode has Closed shadow root type.
PASS: DOMNode is not in a UserAgent shadow tree.

-- Running test case: ShadowRootType.Open
PASS: DOMNode has a single shadow root.
PASS: DOMNode has Open shadow root type.
PASS: DOMNode is not in a UserAgent shadow tree.

-- Running test case: NoShadowRootType
PASS: DOMNode has no shadow roots.
PASS: DOMNode has no shadow root type.
